Test cleanup
Move event tests into the TestServer class
Add missing docstring to tests
